add basic display

+++ src/TestTask.UI/Models/CalculationModel.cs
@@ -2,5 +2,5 @@ namespace TestTask.UI.Models;
 
 public class CalculationModel
 {
-    public decimal InputValue { get; set; }
+    public double InputValue { get; set; }
 }
